Convert Zest scripts in EDT to prevent a deadlock when installing
Prevent a deadlock between EDT and installation thread when installing
the add-on through the marketplace. The deadlock happens when EDT tries
to paint with Zest custom tree cell renderers whose classes were not yet
loaded leading the EDT to access the add-on class loader which is
already being accessed by the installation thread, which ends up waiting
for EDT when converting the plain scripts to Zest scripts.
This change should be considered a band-aid (proper fix would require
core changes to install/add all view components in the EDT).