Issue 1509 - Add "Selenium" add-on
Add "Selenium" add-on that provides WebDrivers and HtmlUnit browser.
Allow to set paths to required WebDrivers and binary both through the
API as the GUI options.
Expose browsers provided by the add-on as well the UI names and a
ComboBoxModel of browsers.
Allow to get WebDrivers with and without a proxy.