Issue 1534 - Change "Ajax Spider" add-on to depend on "Selenium" add-on
Change ZapAddOn.xml file to declare the dependency on "Selenium" add-on,
update version and changes.
Remove the libraries provided by the "Selenium" add-on, add the
add-on to main lib (required for compilation), change to use the classes
provided by the add-on to choose a browser and get a WebDriver.
Change ExtensionAjax to report the active actions and set to depend on
ExtensionSelenium.